Fix lint + SonarCloud: extract ChaosEngineBuilder, deduplicate chaos loop via WriteRowsWithChaosAsync, fix OptWriter whitespace
